backend/requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==41.0.5
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-search-url==0.1
Django==5.0.2
django-cache-url==3.4.4
django-celery-beat==2.7.0
django-ckeditor==6.7.0
django-configurations==2.4.1
django-cors-headers==4.3.0
django-filter==24.2
django-fsm==2.8.1
django-import-export==3.3.1
django-js-asset==2.2.0
django-phonenumber-field==8.1.0
django-prometheus==2.3.1
django-redis==6.0.0
django-silk==5.4.3
django-taggit==5.0.1
django-templated-mail==1.1.1
django-timezone-field==7.0
django-tinymce==3.6.1
django-viewflow==2.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.1
drf-extensions==0.7.1
drf-fsm==1.0.4
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.55.8
fpdf==1.7.2
gprof2dot==2025.4.14
gunicorn==21.2.0
hiredis==3.2.1
idna==3.4
inflection==0.5.1
kiwisolver==1.4.8
kombu==5.3.5
MarkupPy==1.14
matplotlib==3.10.0
msgpack==1.1.1
numpy==2.2.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.3
persiantools==4.2.0
phonenumbers==9.0.15
Pillow==10.1.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
seaborn==0.13.2
setuptools==69.1.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.2
xlwt==1.3.0
